+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 spring.application.name=emoryserver
+spring.data.mongodb.uri=mongodb://localhost:27017/emory
